# server/services/swagger/api-docs.yaml
openapi: 3.0.0
info:
  title: SGL API
  description: API para el Sistema de Gestión de Laboratorios (SGL). Este sistema está diseñado para facilitar la gestión de equipos, usuarios, reservas, y mantenimiento de laboratorios.
  version: 1.0.0
  contact:
    name: Angelo G
    email: angelogarcia2021@@gmail.com
    url: https://wa.me/593982718318
servers:
- url: http://localhost:3000/api
  description: API local
paths:
...
  /role/all:
    get:
      summary: "Obtener todos los roles"
      description: |
        Esta ruta permite obtener todos los roles del sistema con soporte de paginación. Si los roles han sido previamente consultados, se recuperan desde la caché para mejorar el rendimiento.

        **Flujo de obtención:**
        1. El cliente realiza una solicitud para obtener todos los roles, especificando la página y el límite de resultados.
        2. El servidor primero verifica si los datos están almacenados en caché.
        3. Si los datos están en caché, se devuelven directamente al cliente.
        4. Si no están en caché, el servidor consulta la base de datos y luego almacena los resultados en caché.
        5. El servidor devuelve la lista de roles con la información de paginación.

        **Códigos de estado:**
        - **200**: Roles obtenidos exitosamente.
        - **500**: Error en el servidor al obtener los roles.
      operationId: "getAllRoleUsers"
      tags:
      - "Roles"
      security:
      - BearerAuth: []
      parameters:
      - name: "page"
        in: "query"
        description: "Número de página para paginación (predeterminado es 1)."
        required: false
        schema:
          type: integer
          example: 1
      - name: "limit"
        in: "query"
        description: "Número de resultados por página (predeterminado es 10)."
        required: false
        schema:
          type: integer
          example: 10
      responses:
        "200":
          description: "Roles obtenidos exitosamente."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Roles obtenidos exitosamente."
                  data:
                    type: object
                    properties:
                      totalRecords:
                        type: integer
                        example: 25
                      totalPages:
                        type: integer
                        example: 3
                      currentPage:
                        type: integer
                        example: 1
                      recordsPerPage:
                        type: integer
                        example: 10
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id_role:
                              type: string
                              example: "1a2b3c4d"
                            role_name:
                              type: string
                              example: "Administrador"
                            description:
                              type: string
                              example: "Rol con acceso completo"
                            createdAt:
                              type: string
                              format: date-time
                              example: "2024-12-10T11:02:23.426Z"
                            updatedAt:
                              type: string
                              format: date-time
                              example: "2024-12-10T11:02:23.426Z"
        "500":
          description: "Error al obtener los roles."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "¡Ops! Ha ocurrido un error. Por favor, inténtalo de nuevo."
  /role/get-by-id/{id}:
    get:
      summary: "Obtener un rol por ID"
      description: |
        Esta ruta permite obtener un rol específico por su ID. El servidor primero intenta obtener el rol desde la base de datos y si no se encuentra, devuelve un mensaje de error.

        **Códigos de estado:**
        - **200**: Rol obtenido exitosamente.
        - **404**: Rol no encontrado.
        - **500**: Error en el servidor al obtener el rol.
      operationId: "getRoleById"
      tags:
      - "Roles"
      security:
      - BearerAuth: []
      parameters:
      - name: "UUID"
        in: "path"
        required: true
        description: "UUID del rol a obtener"
        schema:
          type: string
          example: "string"
      responses:
        "200":
          description: "Rol obtenido exitosamente."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Rol obtenido exitosamente."
                  data:
                    type: object
                    properties:
                      id_role:
                        type: string
                        example: "1a2b3c4d"
                      role_name:
                        type: string
                        example: "Administrador"
                      description:
                        type: string
                        example: "Rol con acceso completo"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T11:02:23.426Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2024-12-10T11:02:23.426Z"
        "404":
          description: "Rol no encontrado."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Rol no encontrado."
        "500":
          description: "Error al obtener el rol."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "¡Ops! Ha ocurrido un error. Por favor, inténtalo de nuevo."
  # LAB
...